Fix clippy warnings and dead code for CI
- Remove unused import (futures::SinkExt)
- Fix useless .into() conversions on Message::Text
- Fix irrefutable if-let pattern in source registry
- Collapse nested if-let/else blocks per clippy
- Remove unnecessary u64 cast
- Suppress dead_code warning on x_to_time (will be used later)

+++ backend/src/example_data.rs
@@ -53,7 +53,7 @@ fn build_i2c_transaction(name: &str, start_ns: u64, label: &str, color: &str) ->
     // START condition: SCL stays high, SDA drops
     let data_start = start_ns + half_period;
     for i in 0..num_bits {
-        let t = data_start + i as u64 * clock_period_ns;
+        let t = data_start + i * clock_period_ns;
         scl_transitions.push(Transition {
             time_ns: t,
             state: SignalState::Low,
